django_ecommerce/shipping_service/shipping/serializers.py
from rest_framework import serializers
from .models import Shipping

class ShippingSerializer(serializers.ModelSerializer):
    class Meta:
        model = Shipping
        fields = ['id', 'order_id', 'shipping_method', 'shipping_cost', 'tracking_number']