django_ecommerce/shipping_service/requirements.txt
django==3.1.12
djangorestframework==3.14.0
psycopg2-binary==2.9.3
requests==2.28.1